Update PHP versions, Docker aliases, and add new functions
- Updated PHP versions in `base.plugin.bash` to include 8.3 and removed older versions.
- Modified `dclear` function in `docker.plugin.bash` to include a confirmation prompt.
- Added `deit` function in `docker.plugin.bash` for selecting and executing commands in matching containers.
- Added `gprune` function in `git.plugin.bash` to remove local branches that do not exist upstream.
- Updated Docker Compose aliases to use the new `docker compose` command.
- Added new general aliases for `bash-it` commands in `general.aliases.bash`.
- Minor updates to `todo.plugin.bash` for environment variable initialization.
- Added new Git aliases for fetching and pulling in `git.aliases.bash`.

+++ plugins/available/docker.plugin.bash
@@ -5,6 +5,14 @@ function dclear {
   about 'Clear all docker overhead like dangling images, volumes and the likes'
   group 'docker'
 
+  echo "Do you want to move forward with the script? (YES/no)"
+  read choice
+  
+  if [ "$choice" = "no" ]; then
+      echo "Exiting the script."
+      return
+  fi
+
   echo ""
   echo "Pruning: Remove all unused containers, networks, images (both dangling and unreferenced) and volumes"
   docker system prune --all --force
@@ -46,3 +54,41 @@ function docker-bash-root {
   docker exec -it -u root $1 bash
 }
 
+# Function to display matching containers and allow user selection
+function deit {
+    container_query="$1"
+    shift 1
+    container_command="${@:-bash}"  # Default to 'bash' if no command is provided
+
+    # Get the list of matching container names
+    matching_containers=($(docker ps --format '{{.Names}}' | grep "$container_query"))
+    count=${#matching_containers[@]}
+
+    if [ "$count" -eq 0 ]; then
+        echo "No containers found with '$container_query' in the name."
+        return 1
+    elif [ "$count" -eq 1 ]; then
+        container_name="${matching_containers[0]}"
+        echo "Found container: $container_name"
+        echo "Executing: docker exec -it $container_name $container_command"
+        docker exec -it "$container_name" $container_command
+    else
+        echo "Multiple containers found with '$container_query' in the name:"
+        for i in "${!matching_containers[@]}"; do
+            echo "$((i + 1)). ${matching_containers[i]}"
+        done
+
+        while true; do
+            read -p "Choose the number of the container you want to use: " choice
+            if [[ "$choice" =~ ^[0-9]+$ ]] && [ "$choice" -ge 1 ] && [ "$choice" -le "$count" ]; then
+                container_name="${matching_containers[$((choice - 1))]}"
+                echo "Selected container: $container_name"
+                echo "Executing: docker exec -it $container_name $container_command"
+                docker exec -it "$container_name" $container_command
+                break
+            else
+                echo "Invalid choice. Please try again."
+            fi
+        done
+    fi
+}
